separated functions into components and renamed some constants
