+++ fe/src/pages/projects/components/index.js
@@ -1,0 +1,3 @@
+export { default as PageActions } from './pageActions';
+export { default as SetProvider } from './setProvider';
+export { default as PageTitle } from './pageTitle';
